remove dupe code
