+++ model/pred_illuDecomp_layer.py
@@ -1,5 +1,6 @@
 import tensorflow as tf
 
+from pinv import pinv
 
 # am is the albedo map, which has shape (batch, height, width, 3[rgb]) 
 # nm is the sparse normal map, which has shape (batch, height, width, 3[x,y,z])
@@ -58,21 +59,3 @@ def illuDecomp(input, am, nm, gamma, masks):
 
 
 	return tf.reshape(L_SHcoeffs, [-1,27])
-
-
-
-def pinv(A, reltol=1e-6):
-	# compute SVD of input A
-	s, u, v = tf.svd(A)
-
-	# invert s and clear entries lower than reltol*s_max
-	atol = tf.reduce_max(s) * reltol
-	s = tf.boolean_mask(s, s>atol)
-	s_inv = tf.diag(1./s)
-
-	# compute v * s_inv * u_t as psuedo inverse
-	return tf.matmul(v, tf.matmul(s_inv, tf.transpose(u)))
-
-
-
-
