+++ test_demo.py
@@ -7,6 +7,7 @@ import argparse
 from model import SfMNet, lambSH_layer, pred_illuDecomp_layer
 from utils import render_sphere_nm
 from pathlib import Path
+from pinv import pinv
 
 
 parser = argparse.ArgumentParser(description='InverseRenderNet')
@@ -56,18 +57,7 @@ output_folder_path.mkdir(exist_ok=True)
 
 
 
-def pinv(A, reltol=1e-6):
-    """compute pseudo inverse for input matrix"""
-    # compute SVD of input A
-    s, u, v = tf.svd(A)
 
-    # invert s and clear entries lower than reltol*s_max
-    atol = tf.reduce_max(s) * reltol
-    s = tf.boolean_mask(s, s > atol)
-    s_inv = tf.diag(1. / s)
-
-    # compute v * s_inv * u_t as psuedo inverse
-    return tf.matmul(v, tf.matmul(s_inv, tf.transpose(u)))
 
 
 # import ipdb
